# Add an extra test case for file_links_api_spec.rb