- The test case checks the response when the links are tried to be
  created for unknown storage.
- Speed up. On my machine is a huge improvement. Thanks @cbliard for your recommendations!
  - Use shared_let and friends to save time spend in factories.
  - Merge some test cases within the same context.

**Before**:
```
Finished in 2 minutes 17.3 seconds (files took 21.25 seconds to load)
132 examples, 0 failures

[TEST PROF INFO] Time spent in factories: 01:28.528 (61.21% of total time)
```
**After**:
```
Finished in 30.12 seconds (files took 16.82 seconds to load)
131 examples, 0 failures

[TEST PROF INFO] Time spent in factories: 00:01.786 (5.05% of total time)
```

p.s.: @wielinde the specs we thought are absent actually present. We looked into `POST /api/v3/file_links` specs, but we should have checked specs for `POST /api/v3/work_packages/:work_package_id/file_links`.