@cbliard 2023-09-11
If the goal is to save time, when running the tests, it looks like a lot of time is spent in the factories:

```
rspec modules/storages/spec/requests/api/v3/file_links/file_links_api_spec.rb
[...]
Finished in 29 seconds (files took 7.75 seconds to load)
132 examples, 0 failures
[...]
[TEST PROF INFO] Time spent in factories: 00:25.524 (80.7% of total time)
```

To reduce this time, you can rely on `shared_let` to create the objects in database only once. They are then reloaded from database before each test case. Much cheaper than creating on each test.

`current_user` and `project` are good candidates.

When doing so:
```
Finished in 15.28 seconds (files took 7.8 seconds to load)
132 examples, 35 failures

[TEST PROF INFO] Time spent in factories: 00:11.517 (64.13% of total time)
```

There are now failures, because some tests need to be rewritten. For instance:

```ruby
    context 'if user has not sufficient permissions' do
      let(:permissions) { %i(view_work_packages) }

      it_behaves_like 'API V3 collection response', 0, 0, 'FileLink', 'Collection' do
        let(:elements) { [] }
      end
    end
```

This wont work anymore as the permissions are applied to the project/user membership which has already been created. It can then be rewritten as

```
    def remove_permissions(user, *permissions_to_remove)
      user_permissions = RolePermission.joins(role: { members: :principal }).where('users.id': user)
      user_permissions.where(permission: permissions_to_remove).delete_all
      user.reload # clear user's project_role_cache
    end

    context 'if user has not sufficient permissions' do
      before do
        remove_permissions(current_user, :view_file_links)
        get path
      end

      it_behaves_like 'API V3 collection response', 0, 0, 'FileLink', 'Collection' do
        let(:elements) { [] }
      end
    end
```

Also the `get path` has to be repeated because it needs to occur after the permissions have been changed.

These changes can be a pain to implement but the time savings are real.
Here it saved half of the total time (Time spent in factories went from 00:25.524 to 00:11.517).

Another easy saving: when creating a work package through the factory, it also creates a user (the author), a project, a status and a priority, unless they are specified on creation. In this test, `author` and `project` are specified for each work package creation, but not `status` nor `priority`.

To have a default priority and status created automatically, you can use `shared_association_default` like so:
```ruby
  shared_association_default(:priority) { create(:priority) }
  shared_association_default(:status) { create(:status) }
```

This saves 2 seconds: Time spent in factories went from 00:11.517 to 00:09.928

And I did not realize that the 2 work packages instances never change. They can be changed to a `shared_let` as well. This saves 2 more seconds again: Time spent in factories went from 00:09.928 to 00:07.792

And `oauth_application`, `storage`, and `another_storage` can also use `shared_let`. It saves another 3 seconds: Time spent in factories went from 00:07.792 to 00:04.512, for a total time of 8.52 seconds (it was 29 seconds at the beginning).